+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the "amiga-debug" extension will be documented in this file.
 
+## 0.4.0
+
+- don't activate Visual Studio Code on break-on-entry
+- show Explorer after debugging is finished
+- hide GDB output
+- support faster debug output via uaelib trap #87
+
 ## 0.3.0
 
 - added keyboard mapping <kbd>^</kbd> = single step, <kbd>Page-up</kbd> = warp mode